refactor: SseApi 패키지 수정
controller -> api
